feat: checkout design
