+++ storefront/src/app/[countryCode]/(checkout)/checkout/page.tsx
@@ -31,11 +31,13 @@ export default async function Checkout() {
   const customer = await getCustomer()
 
   return (
-    <div className="grid grid-cols-1 small:grid-cols-[1fr_416px] content-container gap-x-40 py-12">
+    <div className="grid grid-cols-1 small:grid-cols-[1fr_416px] content-container gap-2 py-24 h-full">
       <Wrapper cart={cart}>
         <CheckoutForm cart={cart} customer={customer} />
       </Wrapper>
-      <CheckoutSummary cart={cart} />
+      <div className="relative">
+        <CheckoutSummary cart={cart} />
+      </div>
     </div>
   )
 }
