+++ storefront/src/modules/checkout/components/billing-address-form/index.tsx
@@ -3,7 +3,7 @@ import Input from "@modules/common/components/input"
 import CountrySelect from "../country-select"
 import { HttpTypes } from "@medusajs/types"
 
-const BillingAddress = ({ cart }: { cart: HttpTypes.StoreCart | null }) => {
+const BillingAddressForm = ({ cart }: { cart: HttpTypes.StoreCart | null }) => {
   const [formData, setFormData] = useState<any>({})
 
   useEffect(() => {
@@ -53,6 +53,15 @@ const BillingAddress = ({ cart }: { cart: HttpTypes.StoreCart | null }) => {
           data-testid="billing-last-name-input"
         />
         <Input
+          label="Company name"
+          name="billing_address.company"
+          value={formData["billing_address.company"]}
+          onChange={handleChange}
+          autoComplete="organization"
+          data-testid="billing-company-input"
+          colSpan={2}
+        />
+        <Input
           label="Address"
           name="billing_address.address_1"
           autoComplete="address-line1"
@@ -60,62 +69,40 @@ const BillingAddress = ({ cart }: { cart: HttpTypes.StoreCart | null }) => {
           onChange={handleChange}
           required
           data-testid="billing-address-input"
+          colSpan={2}
         />
-        <Input
-          label="Company"
-          name="billing_address.company"
-          value={formData["billing_address.company"]}
-          onChange={handleChange}
-          autoComplete="organization"
-          data-testid="billing-company-input"
-        />
-        <Input
-          label="Postal code"
-          name="billing_address.postal_code"
-          autoComplete="postal-code"
-          value={formData["billing_address.postal_code"]}
-          onChange={handleChange}
-          required
-          data-testid="billing-postal-input"
-        />
-        <Input
-          label="City"
-          name="billing_address.city"
-          autoComplete="address-level2"
-          value={formData["billing_address.city"]}
-          onChange={handleChange}
-          required
-          data-testid="billing-city-input"
-        />
-        <CountrySelect
-          name="billing_address.country_code"
-          autoComplete="country"
-          region={cart?.region}
-          value={formData["billing_address.country_code"]}
-          onChange={handleChange}
-          required
-          data-testid="billing-country-select"
-        />
-        <Input
-          label="State / Province"
-          name="billing_address.province"
-          autoComplete="address-level1"
-          value={formData["billing_address.province"]}
-          onChange={handleChange}
-          required
-          data-testid="billing-province-input"
-        />
-        <Input
-          label="Phone"
-          name="billing_address.phone"
-          autoComplete="tel"
-          value={formData["billing_address.phone"]}
-          onChange={handleChange}
-          data-testid="billing-phone-input"
-        />
+        <div className="grid grid-cols-3 gap-4 col-span-2">
+          <Input
+            label="Postal code"
+            name="billing_address.postal_code"
+            autoComplete="postal-code"
+            value={formData["billing_address.postal_code"]}
+            onChange={handleChange}
+            required
+            data-testid="billing-postal-code-input"
+          />
+          <Input
+            label="City"
+            name="billing_address.city"
+            autoComplete="address-level2"
+            value={formData["billing_address.city"]}
+            onChange={handleChange}
+            required
+            data-testid="billing-city-input"
+          />
+          <CountrySelect
+            name="billing_address.country_code"
+            autoComplete="country"
+            region={cart?.region}
+            value={formData["billing_address.country_code"]}
+            onChange={handleChange}
+            required
+            data-testid="billing-country-select"
+          />
+        </div>
       </div>
     </>
   )
 }
 
-export default BillingAddress
+export default BillingAddressForm
