+++ storefront/src/modules/checkout/components/review/index.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { Heading, Text, clx } from "@medusajs/ui"
+import { Container, Heading, Text, clx } from "@medusajs/ui"
 
 import PaymentButton from "../payment-button"
 import { useSearchParams } from "next/navigation"
@@ -19,36 +19,35 @@ const Review = ({ cart }: { cart: any }) => {
     (cart.payment_collection || paidByGiftcard)
 
   return (
-    <div className="bg-white">
-      <div className="flex flex-row items-center justify-between mb-6">
-        <Heading
-          level="h2"
-          className={clx(
-            "flex flex-row text-3xl-regular gap-x-2 items-baseline",
-            {
+    <Container>
+      <div className="flex flex-col gap-y-2">
+        <div className="flex flex-row items-center justify-between w-full">
+          <Heading
+            level="h2"
+            className={clx("flex flex-row text-xl gap-x-2 items-baseline", {
               "opacity-50 pointer-events-none select-none": !isOpen,
-            }
-          )}
-        >
-          Review
-        </Heading>
-      </div>
-      {isOpen && previousStepsCompleted && (
-        <>
-          <div className="flex items-start gap-x-1 w-full mb-6">
-            <div className="w-full">
-              <Text className="txt-medium-plus text-ui-fg-base mb-1">
-                By clicking the Place Order button, you confirm that you have
-                read, understand and accept our Terms of Use, Terms of Sale and
-                Returns Policy and acknowledge that you have read Medusa
-                Store&apos;s Privacy Policy.
-              </Text>
+            })}
+          >
+            Contact Information
+          </Heading>
+        </div>
+        {isOpen && previousStepsCompleted && (
+          <>
+            <div className="flex items-start gap-x-1 w-full mb-6">
+              <div className="w-full">
+                <Text className="txt-medium-plus text-ui-fg-base mb-1">
+                  By clicking the Place Order button, you confirm that you have
+                  read, understand and accept our Terms of Use, Terms of Sale
+                  and Returns Policy and acknowledge that you have read Medusa
+                  Store&apos;s Privacy Policy.
+                </Text>
+              </div>
             </div>
-          </div>
-          <PaymentButton cart={cart} data-testid="submit-order-button" />
-        </>
-      )}
-    </div>
+            <PaymentButton cart={cart} data-testid="submit-order-button" />
+          </>
+        )}
+      </div>
+    </Container>
   )
 }
