+++ storefront/src/modules/checkout/templates/checkout-form/index.tsx
@@ -1,16 +1,22 @@
 import { listCartShippingMethods } from "@lib/data/fulfillment"
 import { listCartPaymentMethods } from "@lib/data/payment"
 import { HttpTypes } from "@medusajs/types"
-import Addresses from "@modules/checkout/components/addresses"
+import BillingAddress from "@modules/checkout/components/billing-address"
+import ShippingAddress from "@modules/checkout/components/shipping-address"
 import Payment from "@modules/checkout/components/payment"
 import Review from "@modules/checkout/components/review"
 import Shipping from "@modules/checkout/components/shipping"
+import LocalizedClientLink from "@modules/common/components/localized-client-link"
+import UTurnArrowRight from "@modules/common/icons/u-turn-arrow-right"
+import Button from "@modules/common/components/button"
+import Company from "@modules/checkout/components/company"
+import { Company as CompanyType } from "types/global"
 
 export default async function CheckoutForm({
   cart,
   customer,
 }: {
-  cart: HttpTypes.StoreCart | null
+  cart: (HttpTypes.StoreCart & { company: CompanyType }) | null
   customer: HttpTypes.StoreCustomer | null
 }) {
   if (!cart) {
@@ -26,22 +32,28 @@ export default async function CheckoutForm({
 
   return (
     <div>
-      <div className="w-full grid grid-cols-1 gap-y-8">
-        <div>
-          <Addresses cart={cart} customer={customer} />
-        </div>
+      <div className="w-full grid grid-cols-1 gap-y-2">
+        <LocalizedClientLink
+          className="flex items-baseline gap-2 text-sm text-neutral-400 hover:text-neutral-500"
+          href="/cart"
+        >
+          <Button variant="secondary">
+            <UTurnArrowRight />
+            Back to shopping cart
+          </Button>
+        </LocalizedClientLink>
 
-        <div>
-          <Shipping cart={cart} availableShippingMethods={shippingMethods} />
-        </div>
+        {cart?.company && <Company cart={cart} />}
 
-        <div>
-          <Payment cart={cart} availablePaymentMethods={paymentMethods} />
-        </div>
+        <ShippingAddress cart={cart} customer={customer} />
 
-        <div>
-          <Review cart={cart} />
-        </div>
+        <BillingAddress cart={cart} customer={customer} />
+
+        <Shipping cart={cart} availableShippingMethods={shippingMethods} />
+
+        <Payment cart={cart} availablePaymentMethods={paymentMethods} />
+
+        <Review cart={cart} />
       </div>
     </div>
   )
