+++ storefront/src/modules/checkout/templates/checkout-summary/index.tsx
@@ -1,5 +1,4 @@
-import { Heading } from "@medusajs/ui"
-
+import { Container, Heading } from "@medusajs/ui"
 import ItemsPreviewTemplate from "@modules/cart/templates/preview"
 import DiscountCode from "@modules/checkout/components/discount-code"
 import CartTotals from "@modules/common/components/cart-totals"
@@ -7,23 +6,17 @@ import Divider from "@modules/common/components/divider"
 
 const CheckoutSummary = ({ cart }: { cart: any }) => {
   return (
-    <div className="sticky top-0 flex flex-col-reverse small:flex-col gap-y-8 py-8 small:py-0 ">
-      <div className="w-full bg-white flex flex-col">
-        <Divider className="my-6 small:hidden" />
-        <Heading
-          level="h2"
-          className="flex flex-row text-3xl-regular items-baseline"
-        >
-          In your Cart
-        </Heading>
-        <Divider className="my-6" />
-        <CartTotals totals={cart} />
-        <ItemsPreviewTemplate items={cart?.items} />
-        <div className="my-6">
-          <DiscountCode cart={cart} />
-        </div>
+    <Container className="sticky top-2 h-fit w-full flex flex-col small:mt-10">
+      <Heading level="h2" className="flex flex-row text-xl items-baseline">
+        In your Cart
+      </Heading>
+      <Divider className="my-6" />
+      <CartTotals totals={cart} />
+      <ItemsPreviewTemplate items={cart?.items} />
+      <div className="my-6">
+        <DiscountCode cart={cart} />
       </div>
-    </div>
+    </Container>
   )
 }
