+++ storefront/src/modules/common/components/divider/index.tsx
@@ -2,7 +2,10 @@ import { clx } from "@medusajs/ui"
 
 const Divider = ({ className }: { className?: string }) => (
   <div
-    className={clx("h-px w-full border-b border-gray-200 mt-1", className)}
+    className={clx(
+      "h-px border-b border-gray-200 ml-[-1.5rem] w-[calc(100%+3rem)] my-2",
+      className
+    )}
   />
 )
