+++ storefront/src/modules/common/icons/file-plus.tsx
@@ -10,7 +10,7 @@ const FilePlus = ({ size = "12", ...props }: IconProps) => {
       fill="none"
       {...props}
     >
-      <g clip-path="url(#clip0_21037_2683)">
+      <g clipPath="url(#clip0_21037_2683)">
         <path
           d="M10.428 2.867L8.133 0.573C7.77 0.209 7.266 0 6.751 0H3.75C2.233 0 1 1.233 1 2.75V4.844C1.413 4.331 2.042 4 2.75 4C3.91 4 4.868 4.883 4.987 6.012C6.117 6.131 7 7.089 7 8.25C7 9.411 6.117 10.369 4.987 10.488C4.922 11.101 4.606 11.634 4.149 12H8.249C9.766 12 10.999 10.767 10.999 9.25V4.249C10.999 3.728 10.796 3.237 10.427 2.867H10.428ZM7 4V1.585C7.024 1.601 7.051 1.612 7.071 1.633L9.365 3.926C9.386 3.947 9.397 3.975 9.414 3.999H6.999L7 4Z"
           fill="#52525B"
