+++ storefront/src/modules/common/icons/logo.tsx
@@ -10,7 +10,7 @@ const LogoIcon = (props: SVGProps<SVGSVGElement>) => (
     {...props}
   >
     <rect width="20" height="20" rx="6" fill="#18181B" />
-    <g clip-path="url(#clip0_21036_1231)">
+    <g clipPath="url(#clip0_21036_1231)">
       <path
         d="M10.0001 11.5548C10.8592 11.5548 11.5557 10.8584 11.5557 9.99924C11.5557 9.14013 10.8592 8.44369 10.0001 8.44369C9.14103 8.44369 8.44458 9.14013 8.44458 9.99924C8.44458 10.8584 9.14103 11.5548 10.0001 11.5548Z"
         fill="white"
