+++ storefront/src/modules/common/icons/user.tsx
@@ -16,7 +16,7 @@ const User: React.FC<IconProps> = ({
       fill="none"
       {...attributes}
     >
-      <g clip-path="url(#clip0_21037_2692)">
+      <g clipPath="url(#clip0_21037_2692)">
         <path
           d="M6 4.991C7.38071 4.991 8.5 3.87171 8.5 2.491C8.5 1.11029 7.38071 -0.00900078 6 -0.00900078C4.61929 -0.00900078 3.5 1.11029 3.5 2.491C3.5 3.87171 4.61929 4.991 6 4.991Z"
           fill="#52525B"
